refactor: use instance.js to create contract instance
